fogTest implemented into cameraTest + fixed fogTest into mainModel
