+++ public/js/mainModel.js
@@ -257,9 +257,6 @@ function main() {
     const renderer = new THREE.WebGLRenderer({ antialias: true, canvas });
     const baseURL = 'https://storage.googleapis.com/fairgrounds-model/';
 
-    const view1Elem = document.querySelector('#view1');
-    const view2Elem = document.querySelector('#view2');
-
     const loadingDiv = document.createElement('div');
     loadingDiv.style.position = 'absolute';
     loadingDiv.style.top = '50%';
@@ -275,13 +272,12 @@ function main() {
 
     const fov = 55;
     const aspect = 2; // the canvas default
-    const near = 0.9;
+    const near = 0.1;
     const far = 650;
     const camera = new THREE.PerspectiveCamera(fov, aspect, near, far);
     camera.position.set(-53.35, 34.54, 4.64);
 
-    // Create a camera helper for the split view
-    const cameraHelper = new THREE.CameraHelper(camera);
+	const scene = new THREE.Scene();
 
     const gui = new GUI();
     
@@ -306,29 +302,11 @@ function main() {
         }
     }
 
-    const controls = new OrbitControls(camera, view1Elem);
+    const controls = new OrbitControls(camera, canvas);
     controls.target.set(-53.35, 31.54, 4.64);
     controls.minDistance = 2;
     controls.maxDistance = 8;
     controls.update();
-
-    // Second camera for the overview
-    const camera2 = new THREE.PerspectiveCamera(
-        60,  // fov
-        2,   // aspect
-        0.1, // near
-        1000, // far
-    );
-    camera2.position.set(0, 100, 100);
-    camera2.lookAt(0, 0, 0);
-    
-    const controls2 = new OrbitControls(camera2, view2Elem);
-    controls2.target.set(0, 5, 0);
-    controls2.update();
-
-    const scene = new THREE.Scene();
-    scene.background = new THREE.Color('black');
-    scene.add(cameraHelper);
 
     // Store shaders that need to be updated with fogTime
     const shaders = [];
@@ -337,28 +315,19 @@ function main() {
         s.uniforms.fogTime = {value: 0.0};
     };
 
-    // Function to set scissor for split view
-    function setScissorForElement(elem) {
-        const canvasRect = canvas.getBoundingClientRect();
-        const elemRect = elem.getBoundingClientRect();
-        
-        // compute a canvas relative rectangle
-        const right = Math.min(elemRect.right, canvasRect.right) - canvasRect.left;
-        const left = Math.max(0, elemRect.left - canvasRect.left);
-        const bottom = Math.min(elemRect.bottom, canvasRect.bottom) - canvasRect.top;
-        const top = Math.max(0, elemRect.top - canvasRect.top);
-        
-        const width = Math.min(canvasRect.width, right - left);
-        const height = Math.min(canvasRect.height, bottom - top);
-        
-        // setup the scissor to only render to that part of the canvas
-        const positiveYUpBottom = canvasRect.height - bottom;
-        renderer.setScissor(left, positiveYUpBottom, width, height);
-        renderer.setViewport(left, positiveYUpBottom, width, height);
-        
-        // return the aspect
-        return width / height;
-    }
+	//Got this scenegraph dump code from the threejs documentation, super helperful
+	//and it looks great in the console
+	function dumpObject( obj, lines = [], isLast = true, prefix = '' ) {
+		const localPrefix = isLast ? '└─' : '├─';
+		lines.push( `${prefix}${prefix ? localPrefix : ''}${obj.name || '*no-name*'} [${obj.type}]` );
+		const newPrefix = prefix + ( isLast ? '  ' : '│ ' );
+		const lastNdx = obj.children.length - 1;
+		obj.children.forEach( ( child, ndx ) => {
+			const isLast = ndx === lastNdx;
+			dumpObject( child, lines, isLast, newPrefix );
+		} );
+		return lines;
+	}
 
     // Add hemisphere light
     {
@@ -420,6 +389,8 @@ function main() {
             const root = gltf.scene;
             
             // Apply shader modification to all meshes in the scene
+			// Also wanted to note this was done by ChatGPT as well. It was part of the problem
+			// to some degree but not as major, this just helps load the shaders properly
             root.traverse((child) => {
                 if (child.isMesh && child.material) {
                     if (Array.isArray(child.material)) {
@@ -433,6 +404,8 @@ function main() {
             });
             
             scene.add(root);
+			console.log(dumpObject(root).join('\n'));
+			
             controls.update();
         },
         (xhr) => {
@@ -479,39 +452,15 @@ function main() {
             s.uniforms.fogTime.value = totalTime;
         }
 
-        resizeRendererToDisplaySize(renderer);
-    
-        // Turn on the scissor
-        renderer.setScissorTest(true);
-    
-        // Render the original view
-        {
-            const aspect = setScissorForElement(view1Elem);
-        
-            // Adjust the camera for this aspect
-            camera.aspect = aspect;
-            camera.updateProjectionMatrix();
-        
-            // Don't draw the camera helper in the original view
-            cameraHelper.visible = false;
-        
-            // Render
-            renderer.render(scene, camera);
-        }
-    
-        // Render from the 2nd camera
-        {
-            const aspect = setScissorForElement(view2Elem);
-        
-            // Adjust the camera for this aspect
-            camera2.aspect = aspect;
-            camera2.updateProjectionMatrix();
-        
-            // Draw the camera helper in the 2nd view
-            cameraHelper.visible = true;
-        
-            renderer.render(scene, camera2);
-        }
+		if(resizeRendererToDisplaySize(renderer) ) {
+			// Adjust the camera for this aspect
+			const canvas = renderer.domElement;
+			camera.aspect = canvas.clientWidth / canvas.clientHeight;
+			camera.updateProjectionMatrix();
+		}
+	
+		// Render
+		renderer.render(scene, camera);
 
         updateCameraPosition();
         requestAnimationFrame(render);
@@ -521,22 +470,22 @@ function main() {
 
     // Add keyboard controls for camera movement
     {
-        view1Elem.addEventListener('keydown', function(event) {
+        canvas.addEventListener('keydown', function(event) {
             if(event.code == 'KeyW'){
                 camera.translateZ(-5);
             }
         });
-        view1Elem.addEventListener('keydown', function(event) {
+        canvas.addEventListener('keydown', function(event) {
             if(event.code == 'KeyA'){
                 camera.translateX(-5);
             }
         });
-        view1Elem.addEventListener('keydown', function(event) {
+        canvas.addEventListener('keydown', function(event) {
             if(event.code == 'KeyS'){
                 camera.translateZ(5);
             }
         });
-        view1Elem.addEventListener('keydown', function(event) {
+        canvas.addEventListener('keydown', function(event) {
             if(event.code == 'KeyD'){
                 camera.translateX(5);
             }
